adds prettier as an automated and opinionated code formatter
if more than one person contributes to the codebase, there are multiple editors at work, all with their own formatting. everyone has to pay special attention not to commit unrelated changes to their actual contribution. with a fixed format for checked-in code, this should be a thing of the past.